Rename inclusive feature as tectonic
The main idea behind naming the feature as "inclusive" was to hide the
tectonic crate as an internal detail. In case a more suitable all-in
compiler emerged, we could simply change it without modifying the API.

I have decided that it would be better to expose tectonic as an Engine,
this makes is easier to work with e.g. pdflatex and any other future
engine can be just added in the same way.